+++ CHANGELOG.md
@@ -2,3 +2,6 @@
 
 ## 1.0.0 — 2023.05.17
 - Initial release
+## 1.0.1 — 2023.05.18
+- Added a Craft::info log on successful subscription
+- Added an auto-subscription feature on user registration